implementation of Coroutines, used in consultation to the local database and necessary changes.

+++ app/src/main/kotlin/com/fourall/aat/models/User.kt
@@ -8,7 +8,7 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 @Entity(tableName = "user")
 data class User(
-    @PrimaryKey(autoGenerate = true) var id: Int = 0,
+    @PrimaryKey(autoGenerate = true) var id: Long = 0,
     var name: String = "",
     var age: String = ""
 ) : Parcelable
